@@ -19,7 +19,7 @@ publish_to: 'none' # Remove this line if you wish to publish to pub.dev
 version: 1.0.0+1
 
 environment:
-  sdk: '>=2.19.0 <3.0.0'
+  sdk: '>=3.0.2 <4.0.0'
 
 # Dependencies specify other packages that your package needs in order to work.
 # To automatically upgrade your package dependencies to the latest versions
@@ -48,8 +48,8 @@ dependencies:
   flutter_spinkit: ^5.1.0
   custom_timer: ^0.2.3
   qr_flutter: ^4.0.0
-  firebase_auth: ^4.4.0
-  firebase_core: ^2.9.0
+  firebase_auth: ^4.6.2
+  firebase_core: ^2.13.1
   quiver: ^3.0.0
   get_storage: ^2.1.1
   dio: ^5.1.2
