@@ -18,45 +18,42 @@ class RegisterButton extends StatelessWidget {
 
   @override
   Widget build(BuildContext context) {
-    return Padding(
-      padding: EdgeInsets.symmetric(horizontal: 10.w),
-      child: Row(
-        children: [
-          Expanded(
-            child: Container(
-              decoration: BoxDecoration(
-                borderRadius: const BorderRadius.all(Radius.circular(10)),
-                color: color
+    return Row(
+      children: [
+        Expanded(
+          child: Container(
+            decoration: BoxDecoration(
+              borderRadius: const BorderRadius.all(Radius.circular(10)),
+              color: color
+            ),
+            child: TextButton(
+              style: TextButton.styleFrom(
+                shape: const RoundedRectangleBorder(
+                    borderRadius: BorderRadius.all(Radius.circular(10))),
               ),
-              child: TextButton(
-                style: TextButton.styleFrom(
-                  shape: const RoundedRectangleBorder(
-                      borderRadius: BorderRadius.all(Radius.circular(10))),
-                ),
-                onPressed: nav,
-                child: title,
-              ),
+              onPressed: nav,
+              child: title,
             ),
           ),
-          SizedBox(
-            width: 15.h,
-          ),
-          InkWell(
-            onTap:function,
-            child: Container(
-              decoration: const BoxDecoration(
-                  color: ggrey,
-                  borderRadius: BorderRadius.all(Radius.circular(10))),
-              child: Padding(
-                padding:
-                EdgeInsets.symmetric(vertical: 12.h, horizontal: 12.w),
-                child: SvgPicture.asset(
-                    "assets/icon/9.svg"),
-              ),
+        ),
+        SizedBox(
+          width: 15.h,
+        ),
+        InkWell(
+          onTap:function,
+          child: Container(
+            decoration: const BoxDecoration(
+                color: ggrey,
+                borderRadius: BorderRadius.all(Radius.circular(10))),
+            child: Padding(
+              padding:
+              EdgeInsets.symmetric(vertical: 12.h, horizontal: 12.w),
+              child: SvgPicture.asset(
+                  "assets/icon/9.svg"),
             ),
           ),
-        ],
-      ),
+        ),
+      ],
     );
   }
 }
